Fix spelling and use decimal instead of double

+++ Exceptions/InsufficientFundsException.cs
@@ -4,7 +4,7 @@ namespace VendingMachine.Exceptions
 {
     public class InsufficientFundsException : ApiException
     {
-        public InsufficientFundsException(double providedFunds, double cost) :
+        public InsufficientFundsException(decimal providedFunds, decimal cost) :
             base ($"You only provided ${providedFunds:0.00} for a snack that costs ${cost:0.00}.") 
         {
             this.statusCode = 400;
